Move BLR_functions into the blr package
We’ll need to move all code from the "python scripts" directory into src/blr/
so that pip install will find and install it. This moves BLR_functions.py to
a module named blr.utils .
