+++ python scripts/bc_extract.py
@@ -4,7 +4,7 @@
 def main():
 
     global args
-    import BLR_functions as BLR, sys
+    import blr.utils as BLR, sys
 
     #
     # Argument parsing
